requirements.txt
# Risk-prioritized debris selection - 0-1 knapsack QUBO
numpy>=1.20
pandas>=1.3
sgp4>=2.21
dwave-neal>=0.5
